chore: update to laravel framework 10

+++ config/hashids.php
@@ -41,7 +41,7 @@ return [
 
         'main' => [
             'salt' => env('HASH_SALT', 'your-salt-string'),
-            'length' => env('HASH_LENGTH', 18),
+            'length' => (int) env('HASH_LENGTH', 18),
         ],
 
         'alternative' => [
